requirements.txt
cairo
pygame
svgwrite
board
RPi.GPIO
adafruit_seesaw